refactor: move networkmap config
