+++ .env.template
@@ -10,7 +10,7 @@ LOGSTASH_PORT=8080
 LOGSTASH_LEVEL='info'
 
 DATABASE_URL=http://0.0.0.0:8529
-DATABASE_NAME=networkmap
+DATABASE_NAME=configuration
 DATABASE_USER=root
 DATABASE_PASSWORD=
 DATABASE_CERT_PATH=
