--- src/interfaces/iConfig.ts
@@ -1,25 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-import { type RedisConfig } from '@frmscoe/frms-coe-lib/lib/interfaces';
-export interface IConfig {
-  maxCPU: number;
-  redis: RedisConfig;
-  dbURL: string;
-  dbName: string;
-  dbUser: string;
-  dbPassword: string;
-  dbCertPath: string;
-  logger: {
-    logstashHost: string;
-    logstashPort: number;
-    logstashLevel: string;
-  };
-  arangoHost: string;
-  arangoPort: number;
-  functionName: string;
-  apmLogging: boolean;
-  apmSecretToken: string;
-  apmURL: string;
-  nodeEnv: string;
-  cacheTTL: number;
-  sidecarHost: string;
-}
